refactor(services): разделён facade source management
DefaultSourceManagementService оставлен публичной точкой входа для source management flow
feed discovery, preview, command persistence, move flow, normalization policy и summary mapping вынесены в focused collaborators
